fix: funçã o de desincrever com limite de hora

+++ src/back/dancesys/src/main/java/com/dancesys/dancesys/service/ChamadaServiceImpl.java
@@ -1,25 +1,29 @@
 package com.dancesys.dancesys.service;
 
 import com.dancesys.dancesys.entity.Aluno;
+import com.dancesys.dancesys.entity.AulaOcorrencia;
 import com.dancesys.dancesys.entity.Chamada;
 import com.dancesys.dancesys.entity.IdsCompostos.ChamadaId;
 import com.dancesys.dancesys.mapper.ChamadaMapper;
 import com.dancesys.dancesys.repository.ChamadaRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalTime;
 import java.util.List;
 
 @Service
 public class ChamadaServiceImpl implements ChamadaService {
     private final ChamadaRepository chamadaRepository;
     private final AlunoServiceImpl alunoServiceImpl;
+    private final AulaOcorrenciaServiceImpl aulaOcorrenciaServiceImpl;
 
     public ChamadaServiceImpl(
             ChamadaRepository chamadaRepository,
-            AlunoServiceImpl alunoServiceImpl
-    ) {
+            AlunoServiceImpl alunoServiceImpl,
+            AulaOcorrenciaServiceImpl aulaOcorrenciaServiceImpl) {
         this.chamadaRepository = chamadaRepository;
         this.alunoServiceImpl = alunoServiceImpl;
+        this.aulaOcorrenciaServiceImpl = aulaOcorrenciaServiceImpl;
     }
 
     public void gerarChamada(List<Long> alunos, Long idAulaOcorrencia){
@@ -35,7 +39,13 @@ public class ChamadaServiceImpl implements ChamadaService {
     @Override
     public Chamada adicionarAluno(Long idAulaOcorrencia, Long idAluno) throws RuntimeException {
         try{
-            if(chamadaRepository.findByIdAluno_Id(idAluno) !=null){ throw new RuntimeException("você ja esta inscrito nesta aula"); }else{ Aluno aluno = alunoServiceImpl.findById(idAluno); if(aluno.getCreditos() == 0){ throw new RuntimeException("Creditos insuficiente"); }else{
+            if(chamadaRepository.findByIdAluno_Id(idAluno) !=null){
+                throw new RuntimeException("você ja esta inscrito nesta aula");
+            }else{
+                Aluno aluno = alunoServiceImpl.findById(idAluno);
+                if(aluno.getCreditos() == 0) {
+                    throw new RuntimeException("Creditos insuficiente");
+                }else{
                     alunoServiceImpl.dimuirCredito(aluno, 1);
                     return chamadaRepository.save(ChamadaMapper.toEntity(idAulaOcorrencia,idAluno,Chamada.faltante));
                 }
@@ -47,9 +57,15 @@ public class ChamadaServiceImpl implements ChamadaService {
 
     @Override
     public String removerAluno(Long idAluno, Long idAulaOcorrencia){
+        AulaOcorrencia ao = aulaOcorrenciaServiceImpl.buscarPorId(idAulaOcorrencia);
+
+        if(LocalTime.now().isBefore(ao.getIdAula().getHorarioInicio().minusHours(1))){
+            throw new RuntimeException("Você só pode se desincrever em ate 1 hora antes da aula");
+        }
+
+        Aluno aluno = alunoServiceImpl.findById(idAluno);
         ChamadaId id =  new ChamadaId(idAluno,idAulaOcorrencia);
         chamadaRepository.deleteById(id);
-        Aluno aluno = alunoServiceImpl.findById(idAluno);
         alunoServiceImpl.aumentarCredito(aluno,1);
 
         return "Removido com sucesso";
